refactor(core): split ai/personas/analyst.ts into analyst/ submodule
Break the 616-line file into schemas, shared types, run-state tracking,
tool builders, and the run loop, each under 500 lines. analyst.ts is
kept as a pure re-export barrel so every import site, including the
apps/pro deep import of @kansoku/core/ai/personas/analyst, keeps
working unchanged.
